+++ src/classes/Sensors/Sensors.h
@@ -13,7 +13,7 @@ class Sensors{
 
         uint8_t sensor_ranges[TOF_SENSOR_COUNT];
         uint8_t sensor_status[TOF_SENSOR_COUNT];
-        uint8_t getRangeMagnitude();
+        uint16_t getRangeMagnitude();
         int16_t getDifference();
         int16_t getPitch(float& observedAngle);
         uint8_t ranges[TOF_SENSOR_COUNT];
